fix: make engine config experimental
We're not yet sure whether we want to expose these configuration options
going forward. We may also move the recent changes into feature flags,
which would make those changes experimental as well. It makes only sense
to move the entire config settings for this into the experimental
section.
